refactor: SONARCLOUD extract common UI into WinScreenScaffold for victory screens
- Create a new `WinScreenScaffold` composable to encapsulate the shared layout, text styling, and return button used across game-ending screens
